requirements1.txt
asgiref==3.4.1
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.22
botocore==1.35.22
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
distro
distro-info
cryptography
cffi==1.16.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.8.0rc2
disposable-email-domains==0.0.111 # New
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.7.2 # New
django-cors-headers==3.11.0
django-crontab==0.7.1
django-decouple==2.1
django-elasticsearch-dsl
django-environ==0.9.0
django-filter==21.1
django-heroku==0.3.1
django-mptt==0.13.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
elasticsearch
elasticsearch-dsl
et-xmlfile
ghostscript
gunicorn
hashids
hiredis
idna
imageio
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.1
numpy==1.26.4
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.1 # New
pandas==2.2.2
pillow==10.4.0
psycopg2
psycopg2-binary
pycparser
pycryptodome
PyJWT
pyotp==2.9.0 # New
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
redis==5.0.3
referencing==0.35.1
requests
requests-oauthlib
rpds-py
s3transfer
setuptools
simplejson==3.17.6
six==1.16.0
social-auth-app-django
social-auth-core
tzdata
uritemplate
urllib3
Wand
whitenoise
xmltodict
httplib2
hyperlink
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonschema
launchpadlib
lazr.restfulclient
lazr.uri==1.0.6
markdown-it-py
MarkupSafe
oauthlib==3.2.2
PyJWT
pytz
PyYAML
requests==2.31.0
rich==13.7.1
service-identity
setuptools
six==1.16.0
dotenv
MarkupSafe==2.1.1
numpy==1.26.4
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.1 # New
pandas==2.2.2
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
asgiref==3.4.1
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.22
botocore==1.35.22
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.8.0rc2
disposable-email-domains==0.0.111 # New
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.7.2 # New
django-cors-headers==3.11.0
django-crontab==0.7.1
django-decouple==2.1
django-environ==0.9.0
django-filter==21.1
django-mptt==0.13.4
django-storages==1.12.3
drf-spectacular==0.27.2
drf-yasg==1.21.7
oauthlib==3.2.2
openapi-codec==1.3.2
redis==5.0.3
referencing==0.35.1

